pyproject.toml
[project]
name = "mcp-trader"
version = "0.3.0"
description = "A sample MCP server for traders with technical analysis capabilities"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "fastmcp>=2.8.0",
    "numpy~=1.26.4",
    "pandas>=2.3.0",
    "pandas-ta>=0.3.14b",
    "python-dotenv>=1.0.1",
    "ta-lib>=0.6.4",
    "yfinance>=0.2.63",
]
keywords = ["mcp", "trading", "technical-analysis", "finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Office/Business :: Financial :: Investment",
]

[project.urls]
Homepage = "https://github.com/wshobson/mcp-trader"
Repository = "https://github.com/wshobson/mcp-trader"
Issues = "https://github.com/wshobson/mcp-trader/issues"

[[project.authors]]
name = "Seth Hobson"
email = "seth@major7apps.com"

[project.optional-dependencies]
http = ["uvicorn>=0.34.3"]
redis = ["redis>=6.2.0"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.11.13",
    "mypy>=1.16.0",
    "types-aiofiles",
    "types-redis",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-trader = "mcp_trader:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_trader"]

[tool.uv]
constraint-dependencies = [
    # Ensure NumPy 1.x for compatibility
    "numpy<2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["src/mcp_trader/tests"]
python_classes = ["Test*", "*Test"]
python_functions = ["test_*"]
python_files = ["test_*.py", "*_test.py"]

[tool.coverage.run]
source = ["src/mcp_trader"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]